fastapi-backend/requirements.txt
fastapi
uvicorn
httpx
psycopg2-binary
pgvector
sentence-transformers
langchain
langchain-community
langchain-huggingface
requests
python-dotenv
